flask_project/requirements.txt
flask==3.0.2
python-pptx==0.6.23
openai==1.12.0 